% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafinder.R
\name{datafinder}
\alias{datafinder}
\title{Finding good data examples}
\usage{
datafinder(pkg)
}
\arguments{
\item{pkg}{The name of a package (as a character string in quotation marks).}
}
\value{
A data frame. Each row corresponds to a variable in a data set and
indicates the type of variable, the number of missing observations, and
maybe more in the future.
}
\description{
The datafinder function will search a package for data and will return useful
information about that data like how many variables are present, what are their
types, are there missing values, and maybe more in the future.
}
\examples{
datafinder("datasets")

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}
}

