% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_packages.R
\name{discover_packages}
\alias{discover_packages}
\title{Discover all the packages with data that are installed}
\usage{
discover_packages()
}
\value{
A data frame. Each row gives the counts of all the
data frames in each package. If there are no data frames in
the package, an empty data frame will be returned.
}
\description{
The discover_packages function finds the number of data frames in installed
packages.
}
\examples{
discover_packages()

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}

Jack Wilburn \email{jackwilburn@tutanota.com}
}
