% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_package.R
\name{explore_package}
\alias{explore_package}
\title{Explore all the datasets available in a package}
\usage{
explore_package(pkg)
}
\arguments{
\item{pkg}{The name of a package (as a character string in quotation marks).}
}
\value{
A data frame. Each row gives the counts (by class) of all the
variables in each data frame in the package. If there are no data frames in
the package, an empty data frame will be returned.
}
\description{
The explore_package function finds all the data frames in an installed
 package and summarizes the number and classes of variables in each.
}
\examples{
explore_package("datasets")

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}
}

