% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_vars.R
\name{list_vars}
\alias{list_vars}
\title{List all the variables in a data frame along with their classes}
\usage{
list_vars(dataframe)
}
\arguments{
\item{dataframe}{The name of a data frame (as a character string in quotes).}
}
\value{
A data frame. Each row of output represents a variable from the
original data frame passed to the dataframe argument. If the input
dataframe is not actually a data frame, then an empty data frame
will be returned.
}
\description{
The list_vars function lists all the variables in a data frame along
with their classes.
}
\examples{
list_vars("CO2")

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}
}
