% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_dataframe.R
\name{visualize_dataframe}
\alias{visualize_dataframe}
\title{Visualize the variables contained in a data frame.}
\usage{
visualize_dataframe(dataframe, override = FALSE)
}
\arguments{
\item{dataframe}{One data frame (as a character string).}

\item{override}{Defaults to False. Set to true to override imposed limitations
on the number of variables included in the plot}
}
\value{
A plot that plots each variable against another. This function
is dependent on the number of rows and variables
}
\description{
The visualize_dataframe function helps visualize data frames with a reasonable
number of variables
}
\examples{
visualize_dataframe("CO2")

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}

Jack Wilburn \email{jackwilburn@tutanota.com}
}
