% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafinder.R
\name{datafinder}
\alias{datafinder}
\title{Catalog the variables in available data sets}
\usage{
datafinder(package, summary = TRUE)
}
\arguments{
\item{package}{The name of a package (as a character string in quotation marks).}

\item{summary}{If summary is TRUE, each row of output will be a single dataframe,
along with the number of each type of variable present. If summary is
FALSE, each row of output will be a variable along with its type.}
}
\value{
A data frame. If summary is TRUE, each row of output will be a
    single dataframe, along with the number of each type of variable
    present. If summary is FALSE, each row of output will be a variable
    along with its type.
}
\description{
The datafinder function will search a package for data frames and will
return useful information about the variables and their types.
Hopefully there will be more information available in the future.
(For example: Is there missing data? How many factors are present?
Are numerical variables normally distributed?)
}
\examples{
datafinder("datasets")

}
\author{
Sean Raleigh, \email{sraleigh@westminstercollege.edu}
}

