% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafinder.R
\docType{package}
\name{datafinder}
\alias{datafinder}
\alias{datafinder-package}
\title{datafinder: a package for cataloging the variables in installed data sets.}
\description{
The datafinder package is designed to help users explore the number
and types of variables available in installed packages. This will
be especially useful for statistics educators who are always on the lookout
for example data sets that have the features needed that best illustrate
the topic at hand.
}

